SetT implementation and test

+++ src/main/kotlin/cpu/RegisterBank.kt
@@ -4,7 +4,7 @@ package cpu
 class RegisterBank {
     private val general = UByteArray(8) {0u}
     val address = UByteArray(2) {0u}
-    var timer = 0u
+    var timer = 0u.toUByte()
 
     fun writeRegister(address: UByte, value: UByte){
         general[address.toInt()] = value
